Move roarmap-rules.js to lib/rules.js
While there, update .gitignore.
