+++ .gitignore
@@ -3,3 +3,4 @@ cache/
 node_modules
 npm-debug.log
 roarmap-dump.json
+test-vector.json.new
