+++ lib/rules.js
